moved utils folder outside of components folder
